+++ src/components/flex/flex.stories.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Flex, FlexProps } from '.';
+import { Image } from '../image';
 import { Story, Meta } from '@storybook/react';
 
 export default {
@@ -11,7 +12,23 @@ export default {
   },
 } as Meta;
 
-const Template: Story<FlexProps> = (args) => <Flex {...args} />;
+const Template: Story<FlexProps> = (args) => (
+  <Flex {...args} justify='space-between'>
+    <Image
+      width='49%'
+      imageUrl='https://source.unsplash.com/V_WBdOLOgJU/600x500'
+      imageAlt='woman in orange knit cap and blue jacket'
+      loading='eager'
+    />
+
+    <Image
+      width='49%'
+      imageUrl='https://source.unsplash.com/V_WBdOLOgJU/600x500'
+      imageAlt='woman in orange knit cap and blue jacket'
+      loading='lazy'
+    />
+  </Flex>
+);
 
 // Base default flex
 export const Base = Template.bind({});